backend/requirements.txt
alembic==1.13.3
asyncpg==0.30.0
fastapi==0.115.6
greenlet==3.1.1
httpx==0.28.1
pydantic-settings==2.7.1
redis==5.2.1
SQLAlchemy==2.0.36
uvicorn[standard]==0.34.0
